+++ CHANGELOG.md
@@ -13,12 +13,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add new converter input format "EdgeList" with EdgeListDecoder. Format has nodes and edges on separate lines, is smaller and faster to convert.
+
 - Breaking. Added version checks for binary data. Requires to convert graph data or add v1 at the top of meta files.
 
 - Add migrate script to pull to new version of deepgnn.
 
+- Add debug mode to MultiWorkersConverter, using debug=True will now disable multiprocessing and show error messages.
+
 ### Changed
-- Remove FeatureType enum, replace with np.dtype. FeatureType.BINARY -> np.uint8, FeatureType.FLOAT -> np.float32, FeatureType.INT64 -> np.int64.
+- Breaking. Remove FeatureType enum, replace with np.dtype. FeatureType.BINARY -> np.uint8, FeatureType.FLOAT -> np.float32, FeatureType.INT64 -> np.int64.
+
+- Rename function deepgnn.graph_engine.data.to_json_node -> deepgnn.graph_engine.data.to_edge_list_node and update functionality accordingly.
 
 ## [0.1.55] - 2022-08-26